Fix duplication
